✨ Added the post deletion.
By deleting a post by using the _PostAPI_ we also delete its reference in
the thread that contains it since we don't want to have a ghost post in a
thread.

We get this container thread by using the *ThreadID* present in a Post. We then
go through all the *PostsID* and *delete* the one that possesses the *same
ID* that the Post we firstly intend to *delete*.
Finally, we *save* these changes in the thread otherwise this deletion would
not be effective.